--- lib/css/index.css
@@ -1,5 +1,0 @@
-@charset "utf-8";
-
-@import "variables.css";
-@import "font.css";
-@import "snackbar.css";
