+++ lib/global.css
@@ -1,0 +1,4 @@
+@charset "utf-8";
+
+@import "globals/colours.css";
+@import "globals/font.css";
