+++ lib/globals/colours.css
@@ -5,6 +5,7 @@
   --error: #ff6238;
   --success: #10c443;
   --warning: #ffd238;
+
   --text-snackbar: #3e3e3e;
 
   --shadow-standard: var(--black-30) 0px 0px 3px 0px;
