+++ web/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Storybook: KatiKati UI Elements</title>
-    <link rel="stylesheet" type='text/css' href="packages/katikati_ui_lib/css/index.css">
+    <link rel="stylesheet" type='text/css' href="packages/katikati_ui_lib/global.css">
+    <link rel="stylesheet" type='text/css' href="packages/katikati_ui_lib/components/snackbar/snackbar.css">
     <script src="script.dart.js" defer></script>
 </head>
 <body>
